[RFC] a naive "unneeded import" check implementation
It seems to work at least in some cases, I still need to check
if it returns false positives, since they'll be annoying.

I did try it on the ldclint sources itself, found and fixed a
couple of issues.

There are still two open issues though:

1. Re-exported symbols. Currently if a symbol is defined in a
   module `A`, which is publicly imported by module `B` and
   your code imports `B` and uses that symbol, the lint will
   report that `B` is an unused import. That's pretty bad and
   this is why I made it disabled by default. Unfortunately
   I don't think this issue is fixable with the current design,
   since it knows nothing about re-exported symbols. Instead of
   checking the definition module of used symbols, we need to
   check the module that brought the symbol to scope. But I
   haven't found an easy way to implement this.
2. There is a weird bug: if I check a single file, an unused
   import is reported correctly, but if I check all files at
   once, the warning disappears. I thought that this is because
   `Context` gets polluted by other modules, so we end up
   thinking an import is actually used, but my attempt to fix
   it (see the next commit) failed.

I also not completely sure:
 - if it's necessary to visit all the nodes I currently visit
 - if there are more nodes that need to be visited ;)

Ideally, I'd love the lint infrastructure to only visit
`Import` and `Dsymbol` for this kind of lint, but now it doesn't
seem to be possible, since not all symbols are visited.

Signed-off-by: Ilya Yanok <ilya.yanok@weka.io>